fix(#7539): extract Emit.Savepoint into its own file to satisfy qulice
